fix: updating package for tests.
