+++ {{ cookiecutter.__package_name_kebab_case }}/__tests__/test_api.py
@@ -3,7 +3,7 @@
 import httpx
 from fastapi.testclient import TestClient
 
-from {{ cookiecutter.__package_name_snake_case }}.serve.api import app
+from serve.api import app
 
 client = TestClient(app)
